temp changes

+++ integration-tests/src/test/java/org/apache/druid/tests/indexer/AbstractITRealtimeIndexTaskTest.java
@@ -191,5 +191,5 @@ public abstract class AbstractITRealtimeIndexTaskTest extends AbstractIndexerTes
 
   abstract void postEvents() throws Exception;
 
-  abstract int getNumExpectedRowsIngested();
+  abstract long getNumExpectedRowsIngested();
 }
